+++ framework/domain-db/src/main/java/module-info.java
@@ -19,7 +19,7 @@
 /**
  * Framework database domain model classes, such as:<br>
  * <br>
- * {@link is.codion.framework.domain.db.DatabaseDomain}
+ * {@link is.codion.framework.domain.db.SchemaDomain}
  */
 module is.codion.framework.domain.db {
 	requires transitive is.codion.framework.domain;