+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.10
 ### is.codion.framework.domain.db
 - Module is.codion.framework.domain.db split from is.codion.tools.generator.domain.
+- DatabaseDomain renamed SchemaDomain.
 
 ## 0.18.9
 ### is.codion